requirements.txt
pandas
requests
sqlalchemy
pyarrow
streamlit>=1.40.0
plotly
schedule
altair<5
python-dotenv
